Convert MultiReport/Keyboard to Dispatcher/Report model
This replaces the single `Keyboard` class with two classes: `Dispatcher` and `Report`. The
`Report` class simply holds the report data (and provides functions to interpret and
manipulate that data), and the `Dispatcher` does the actual HID communication. This allows
us to pass the report data to plugins without them needing access to the dispatcher
itself. It also means we only need to store a single set of keyboard report data in memory
at all times.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>