Fix HID keyboard boot protocol
I'm not really clear on how this works, but with this change, I use a separate
variabl (`boot_protocol_`) to determine which type of report to send, and let
`hid_protocol_` be always set to `nkro_mode`. Then, when PluggableUSB() calls the
`setup()` method of the dispatcher, it seems like it tells the host that it's in NKRO
protocol, but sends 6KRO reports anyway, and that appears to work on both macOS & Linux.

There is still a bug where keys can get stuck on if I switch protocol while holding those
keys, however.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>